feat(redis): fallback to inmemory cache and support env interpolation (#3553)
* feat(redis): fallback to inmemory cache and support env interpolation

* Add tsOnly param to build

* Remove unnecessary stuff

* feat(cli): ignore ts errors in artifacts

* Automatic dotenv, ts-node and tsconfig-paths integration

* feat(json-schema/openapi/raml): respect error status codes

* Handle responses correctlyt

* enhance(json-machete): do not look for annotated schemas

* No need to create union for single response result

* feat(raml): respect trait definitions

* Improve naming

* Respect existing titles

* Correct error message

* Add ignoreErrorResponses flag

* Expose MeshContext with MeshInstance

* Respect higher level traits

* No more parse additionalTypeDefs

* Fix TS

* Update custom server example

* Support tsnode even if tsconfig doesn't exist

* ignoreErrorResponses for OpenAPI loader

* Fix Redis tests

+++ packages/cli/src/commands/ts-artifacts.ts
@@ -49,7 +49,9 @@ function buildSignatureBasedOnRootFields(
       },
     };
 
-    operationMap[fieldName] = `  ${field.name}: InContextSdkMethod<${codegenHelpers.getTypeToUse(
+    operationMap[fieldName] = `  /** ${field.description} **/\n  ${
+      field.name
+    }: InContextSdkMethod<${codegenHelpers.getTypeToUse(
       parentTypeNode
     )}['${fieldName}'], ${argsName}, ${unifiedContextIdentifier}>`;
   }
@@ -108,6 +110,7 @@ export async function generateTsArtifacts({
   meshConfigCode,
   logger,
   sdkConfig,
+  tsOnly = false,
 }: {
   unifiedSchema: GraphQLSchema;
   rawSources: RawSourceOutput[];
@@ -119,6 +122,7 @@ export async function generateTsArtifacts({
   meshConfigCode: string;
   logger: Logger;
   sdkConfig: YamlConfig.SDKConfig;
+  tsOnly: boolean;
 }) {
   const artifactsDir = join(baseDir, '.mesh');
   logger.info('Generating index file in TypeScript');
@@ -238,9 +242,9 @@ export const documentsInSDL = /*#__PURE__*/ [${documents.map(
             documentSource => `/* GraphQL */\`${documentSource.rawSDL}\``
           )}];
 
-export async function getBuiltMesh(): Promise<MeshInstance> {
+export async function getBuiltMesh(): Promise<MeshInstance<MeshContext>> {
   const meshConfig = await getMeshOptions();
-  return getMesh(meshConfig);
+  return getMesh<MeshContext>(meshConfig);
 }
 
 export async function getMeshSDK<TGlobalContext = any, TGlobalRoot = any, TOperationContext = any, TOperationRoot = any>(sdkOptions?: SdkOptions<TGlobalContext, TGlobalRoot>) {
@@ -288,45 +292,47 @@ export async function getMeshSDK<TGlobalContext = any, TGlobalRoot = any, TOpera
   const tsFilePath = join(artifactsDir, 'index.ts');
   await writeFile(tsFilePath, codegenOutput.replace(BASEDIR_ASSIGNMENT_COMMENT, baseUrlAssignmentESM));
 
-  logger.info('Compiling TS file as ES Module to `index.mjs`');
-  const jsFilePath = join(artifactsDir, 'index.js');
-  const dtsFilePath = join(artifactsDir, 'index.d.ts');
-  compileTS(tsFilePath, ts.ModuleKind.ESNext, [jsFilePath, dtsFilePath]);
+  if (!tsOnly) {
+    logger.info('Compiling TS file as ES Module to `index.mjs`');
+    const jsFilePath = join(artifactsDir, 'index.js');
+    const dtsFilePath = join(artifactsDir, 'index.d.ts');
+    compileTS(tsFilePath, ts.ModuleKind.ESNext, [jsFilePath, dtsFilePath]);
 
-  const mjsFilePath = join(artifactsDir, 'index.mjs');
-  await rename(jsFilePath, mjsFilePath);
+    const mjsFilePath = join(artifactsDir, 'index.mjs');
+    await rename(jsFilePath, mjsFilePath);
 
-  logger.info('Writing index.ts for CJS to the disk.');
-  await writeFile(tsFilePath, codegenOutput.replace(BASEDIR_ASSIGNMENT_COMMENT, baseUrlAssignmentCJS));
+    logger.info('Writing index.ts for CJS to the disk.');
+    await writeFile(tsFilePath, codegenOutput.replace(BASEDIR_ASSIGNMENT_COMMENT, baseUrlAssignmentCJS));
 
-  logger.info('Compiling TS file as CommonJS Module to `index.js`');
-  compileTS(tsFilePath, ts.ModuleKind.CommonJS, [jsFilePath, dtsFilePath]);
+    logger.info('Compiling TS file as CommonJS Module to `index.js`');
+    compileTS(tsFilePath, ts.ModuleKind.CommonJS, [jsFilePath, dtsFilePath]);
 
-  await writeJSON(join(artifactsDir, 'package.json'), {
-    name: 'mesh-artifacts',
-    private: true,
-    type: 'commonjs',
-    main: 'index.js',
-    module: 'index.mjs',
-    sideEffects: false,
-    typings: 'index.d.ts',
-    typescript: {
-      definition: 'index.d.ts',
-    },
-    exports: {
-      '.': {
-        require: './index.js',
-        import: './index.mjs',
+    await writeJSON(join(artifactsDir, 'package.json'), {
+      name: 'mesh-artifacts',
+      private: true,
+      type: 'commonjs',
+      main: 'index.js',
+      module: 'index.mjs',
+      sideEffects: false,
+      typings: 'index.d.ts',
+      typescript: {
+        definition: 'index.d.ts',
       },
-      './*': {
-        require: './*.js',
-        import: './*.mjs',
+      exports: {
+        '.': {
+          require: './index.js',
+          import: './index.mjs',
+        },
+        './*': {
+          require: './*.js',
+          import: './*.mjs',
+        },
       },
-    },
-  });
+    });
 
-  logger.info('Deleting index.ts');
-  await unlink(tsFilePath);
+    logger.info('Deleting index.ts');
+    await unlink(tsFilePath);
+  }
 }
 
 export function compileTS(tsFilePath: string, module: ts.ModuleKind, outputFilePaths: string[]) {
